Restructure encoder modules
- The top-level depparse module is renamed to encoder::deprel.
- The Layer{Encoder,Decoder} are moved to the encoder::layer module.
