+++ sticker/src/encoder/deprel/relative_position.rs
@@ -5,7 +5,7 @@ use serde_derive::{Deserialize, Serialize};
 use super::{
     attach_orphans, break_cycles, find_or_create_root, DecodeError, DependencyEncoding, EncodeError,
 };
-use crate::{EncodingProb, SentenceDecoder, SentenceEncoder};
+use crate::encoder::{EncodingProb, SentenceDecoder, SentenceEncoder};
 
 /// Relative head position.
 ///
@@ -119,8 +119,8 @@ mod tests {
     use conllx::token::TokenBuilder;
 
     use super::{RelativePosition, RelativePositionEncoder};
-    use crate::depparse::{DecodeError, DependencyEncoding};
-    use crate::{EncodingProb, SentenceDecoder};
+    use crate::encoder::deprel::{DecodeError, DependencyEncoding};
+    use crate::encoder::{EncodingProb, SentenceDecoder};
 
     // Small tests for the relative position encoder. Automatic
     // testing is performed in the module tests.
