+++ sticker/src/encoder/mod.rs
@@ -1,10 +1,16 @@
+//! Label encoders.
+
 use std::borrow::{Borrow, Cow};
 use std::hash::Hash;
 
-use conllx::graph::{Node, Sentence};
-use failure::{format_err, Error};
+use conllx::graph::Sentence;
+use failure::Error;
 
-use crate::{Layer, LayerValue, Numberer};
+use crate::Numberer;
+
+pub mod deprel;
+
+pub mod layer;
 
 /// An encoding with its probability.
 pub struct EncodingProb<'a, E>
@@ -78,63 +84,6 @@ pub trait SentenceEncoder {
 
     /// Encode the given sentence.
     fn encode(&mut self, sentence: &Sentence) -> Result<Vec<Self::Encoding>, Error>;
-}
-
-/// Encode sentences using a CoNLL-X layer.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct LayerEncoder {
-    layer: Layer,
-}
-
-impl LayerEncoder {
-    /// Construct a new layer encoder of the given layer.
-    pub fn new(layer: Layer) -> Self {
-        LayerEncoder { layer }
-    }
-}
-
-impl SentenceDecoder for LayerEncoder {
-    type Encoding = String;
-
-    fn decode<'a, S>(&self, labels: &[S], sentence: &mut Sentence) -> Result<(), Error>
-    where
-        S: AsRef<[EncodingProb<'a, Self::Encoding>]>,
-        Self::Encoding: 'a,
-    {
-        assert_eq!(
-            labels.len(),
-            sentence.len() - 1,
-            "Labels and sentence length mismatch"
-        );
-
-        for (token, token_labels) in sentence
-            .iter_mut()
-            .filter_map(Node::token_mut)
-            .zip(labels.iter())
-        {
-            if let Some(label) = token_labels.as_ref().get(0) {
-                token.set_value(&self.layer, label.encoding().as_str());
-            }
-        }
-
-        Ok(())
-    }
-}
-
-impl SentenceEncoder for LayerEncoder {
-    type Encoding = String;
-
-    fn encode(&mut self, sentence: &Sentence) -> Result<Vec<Self::Encoding>, Error> {
-        let mut encoding = Vec::with_capacity(sentence.len() - 1);
-        for token in sentence.iter().filter_map(Node::token) {
-            let label = token
-                .value(&self.layer)
-                .ok_or_else(|| format_err!("Token without a label: {}", token.form()))?;
-            encoding.push(label.to_owned());
-        }
-
-        Ok(encoding)
-    }
 }
 
 /// An encoder wrapper that encodes/decodes to a categorical label.
@@ -213,8 +162,9 @@ mod tests {
 
     use conllx::io::Reader;
 
-    use super::{CategoricalEncoder, LayerEncoder};
-    use crate::{EncodingProb, Layer, Numberer, SentenceDecoder, SentenceEncoder};
+    use super::layer::LayerEncoder;
+    use super::{CategoricalEncoder, EncodingProb, SentenceDecoder, SentenceEncoder};
+    use crate::{Layer, Numberer};
 
     static NON_PROJECTIVE_DATA: &'static str = "testdata/nonprojective.conll";
