+++ sticker/src/lib.rs
@@ -1,12 +1,7 @@
 mod collector;
 pub use crate::collector::{Collector, NoopCollector};
 
-mod encoder;
-pub use crate::encoder::{
-    CategoricalEncoder, EncodingProb, LayerEncoder, SentenceDecoder, SentenceEncoder,
-};
-
-pub mod depparse;
+pub mod encoder;
 
 mod input;
 pub use crate::input::{Embeddings, LayerEmbeddings, SentVectorizer};
